Move caches to separate package
